moved users page inside profile page; also changed the profile layout
