+++ client/src/layouts/admin/types.d.ts
@@ -1,6 +1,8 @@
 import type React from 'react';
+import type { TabsType } from 'components/tabs';
 
 export interface AdminLayoutProps extends React.PropsWithChildren {
   children: React.ReactNode;
   title?: string;
+  adminTabs?: TabsType;
 }
