+++ client/src/layouts/profile/constants.ts
@@ -1,0 +1,12 @@
+import type { TabsType } from 'components/tabs';
+
+export const USER_TABS: TabsType = {
+  PROFILE: {
+    name: 'Profile',
+    href: '/profile',
+  },
+  USERS: {
+    name: 'Users',
+    href: '/profile/users',
+  },
+};
